+++ src/internal/main.h
@@ -21,7 +21,7 @@ baton_result_t parse_toggle_bitfield_cmd( int fd, yuck_t *argp );
 baton_result_t parse_read_bitfield_cmd( int fd, yuck_t *argp );
 baton_result_t check_relay_argument( int * const relay_num, yuck_t * const argp );
 baton_result_t check_id_argument( char * const id, yuck_t * const argp );
-baton_result_t check_bitfield_argument( unsigned int * const bitfield, yuck_t * const argp );
+baton_result_t check_bitfield_argument( unsigned long * const bitfield, yuck_t * const argp );
 
 
 #endif /* _INTERNAL_BATON_H_ */
